requirements/prod.txt
aio_pika
fastapi
pydantic
pyjwt
uvicorn
websockets
